Renamed enum to not interfere with regular types

+++ include/gotchangpdf/c_page_tree.h
@@ -9,7 +9,7 @@ extern "C"
 {
 #endif
 
-	GOTCHANG_PDF_API long long CALLING_CONVENTION PageTree_GetPageCount(PageTreeHandle);
+	GOTCHANG_PDF_API int CALLING_CONVENTION PageTree_GetPageCount(PageTreeHandle);
 	GOTCHANG_PDF_API PageObjectHandle CALLING_CONVENTION PageTree_GetPage(PageTreeHandle, int at);
 	GOTCHANG_PDF_API void CALLING_CONVENTION PageTree_Release(PageTreeHandle);
